+++ edison-jobs/src/main/java/de/otto/edison/jobs/repository/JobRepository.java
@@ -42,4 +42,5 @@ public interface JobRepository {
 
     long size();
 
+    void deleteAll();
 }